+++ world/visibility/fov.cpp
@@ -1,4 +1,6 @@
+#include <algorithm>
 #include <cmath>
+#include <memory>
 #include "fov.h"
 
 const int Fov::Multipliers[4][8] = {
@@ -11,26 +13,26 @@ const int Fov::Multipliers[4][8] = {
 void Fov::cast_light(
     std::shared_ptr<iFov>& fov_response,
     std::shared_ptr<iWallMap>& map,
-    uint row,
-    uint cell,
-    uint radius,
-    uint row_start,
+    unsigned int row,
+    unsigned int cell,
+    unsigned int radius,
+    unsigned int row_start,
     float start_slope,
     float end_slope,
-    uint xx,
-    uint xy,
-    uint yx,
-    uint yy)
+    int xx,
+    int xy,
+    int yx,
+    int yy)
 {
     if (start_slope < end_slope)
     {
         return;
     }
     float next_start_slope = start_slope;
-    for (uint i = row_start; i <= radius; i++)
+    for (unsigned int i = row_start; i <= radius; i++)
     {
         bool blocked = false;
-        for (int dx = -i, dy = -i; dx <= 0; dx++)
+        for (int dx = -(int)i, dy = -(int)i; dx <= 0; dx++)
         {
             float l_slope = static_cast<float>((dx - 0.5) / (dy + 0.5));
             float r_slope = static_cast<float>((dx + 0.5) / (dy - 0.5));
@@ -45,20 +47,20 @@ void Fov::cast_light(
 
             int sax = dx * xx + dy * xy;
             int say = dx * yx + dy * yy;
-            if ((sax < 0 && (uint)std::abs(sax) > cell) ||
-                (say < 0 && (uint)std::abs(say) > row))
+            if ((sax < 0 && (unsigned int)std::abs(sax) > cell) ||
+                (say < 0 && (unsigned int)std::abs(say) > row))
             {
                 continue;
             }
-            uint ax = cell + sax;
-            uint ay = row + say;
+            unsigned int ax = cell + (unsigned int)sax;
+            unsigned int ay = row + (unsigned int)say;
             if (ax >= map->width() || ay >= map->height())
             {
                 continue;
             }
 
-            uint radius2 = radius * radius;
-            if ((uint)(dx * dx + dy * dy) < radius2)
+            unsigned int radius2 = radius * radius;
+            if ((unsigned int)(dx * dx + dy * dy) < radius2)
             {
                 // todo: save the square of the distance
                 fov_response->set_fov(ay, ax, static_cast<unsigned int>(sqrt(dx * dx + dy * dy)));
@@ -107,20 +109,20 @@ void Fov::cast_light(
 void Fov::do_fov(
     std::shared_ptr<iFov>& fov_response,
     std::shared_ptr<iWallMap>& map,
-    uint row,
-    uint cell)
+    unsigned int row,
+    unsigned int cell)
 {
     // slightly larger than the actual maximum (diagonal)
-    int longest_dist = map->height() + map->width(); 
+    unsigned int longest_dist = map->height() + map->width(); 
     Fov::do_fov(fov_response, map, row, cell, longest_dist);    
 }
 
 void Fov::do_fov(
     std::shared_ptr<iFov>& fov_response,
     std::shared_ptr<iWallMap>& map,
-    uint row,
-    uint cell,
-    uint radius)
+    unsigned int row,
+    unsigned int cell,
+    unsigned int radius)
 {
     fov_response->set_fov(row, cell, 0);
     if (radius == 0)
@@ -128,7 +130,7 @@ void Fov::do_fov(
         radius = std::max(map->height(), map->width());
     }
 
-    for (uint i = 0; i < 8; i++)
+    for (unsigned int i = 0; i < 8; i++)
     {
         cast_light(
             fov_response,
